chore: move computed impl to separate file
